refactor: move PLPChunkWriter into plp module
