+++ tiberius/src/protocol.rs
@@ -624,47 +624,3 @@ impl<'a, I: Io> Write for PacketWriter<'a, I> {
         Ok(())
     }
 }
-
-pub struct PLPChunkWriter<W: Write> {
-    pub target: W,
-    pub buf: Vec<u8>,
-}
-
-impl<W: Write> io::Write for PLPChunkWriter<W> {
-    fn write(&mut self, buf: &[u8]) -> io::Result<usize> {
-        // fast path for small writes
-        if self.buf.capacity() - self.buf.len() > buf.len() {
-            self.buf.extend_from_slice(buf);
-            return Ok(buf.len());
-        }
-
-        let mut pending = buf;
-        while !pending.is_empty() {
-            let free_bytes = self.buf.capacity() - self.buf.len();
-            let boundary = cmp::min(pending.len(), free_bytes);
-            let (fitting, next) = pending.split_at(boundary);
-
-            // we can produce a whole chunk => write to the underlying buf
-            if fitting.len() == free_bytes {
-                self.target
-                    .write_u32::<LittleEndian>(self.buf.capacity() as u32)?;
-                self.target.write_all(&self.buf)?;
-                self.target.write_all(fitting)?;
-                self.buf.truncate(0);
-            } else {
-                self.buf.extend_from_slice(fitting);
-            }
-            pending = next;
-        }
-        Ok(buf.len())
-    }
-
-    fn flush(&mut self) -> io::Result<()> {
-        if !self.buf.is_empty() {
-            self.target.write_u32::<LittleEndian>(self.buf.len() as u32)?;
-            self.target.write_all(&self.buf)?;
-            self.buf.truncate(0);
-        }
-        Ok(())
-    }
-}
